requirements.txt
pandas==1.5.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest==7.3.1
pytest-cov==4.1.0